% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getHighchartsColors}
\alias{getHighchartsColors}
\title{Get Highcharts Default Colors}
\usage{
getHighchartsColors()
}
\description{
Creates a vector of the default Highcharts colors
as of v4.0.3
}
\seealso{
Other shinyhighcharts.elements: \code{\link{JSONify}};
  \code{\link{highchartsOutput}};
  \code{\link{renderHighcharts}}
}

