% Generated by roxygen2 (4.0.1): do not edit by hand
\name{highchartsOutput}
\alias{highchartsOutput}
\title{Highcharts Heatmap Output}
\usage{
highchartsOutput(inputId, width = "100\%", height = "400px",
  type = c("heatmap"))
}
\arguments{
\item{inputId}{}

\item{width}{}

\item{height}{}

\item{type}{A character vector indicating the type of highchart to output}
}
\description{
Render a renderHighHeatmap within an application page.
}
\seealso{
Other shinyhighcharts.elements: \code{\link{renderHighHeatmap}}
}

