% Generated by roxygen2 (4.0.1): do not edit by hand
\name{JSONify}
\alias{JSONify}
\title{Convert a Data Frame into a JSON Object}
\usage{
JSONify(df, element.names = NULL)
}
\description{
Converts the rows of a data frame into a list of vectors.
Optionally, include names for each element.
}
\seealso{
Other shinyhighcharts.elements: \code{\link{getHighchartsColors}};
  \code{\link{highchartsOutput}};
  \code{\link{renderHighcharts}}
}

