% Generated by roxygen2 (4.0.1): do not edit by hand
\name{renderHighBoxplot}
\alias{renderHighBoxplot}
\title{Render Highcharts Boxplot}
\usage{
renderHighBoxplot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that evaluates to a matrix or dataframe object
with three columns.  The first column is the y axis value, the second
is the x axis value, and the third column is the color value for the
quadrant.}

\item{env}{}

\item{quoted}{}
}
\description{
Creates a Highcharts boxplot output from a matrix of dataframe.
}
\seealso{
Other shinyhighcharts.elements: \code{\link{highchartsOutput}};
  \code{\link{renderHighColumnChart}};
  \code{\link{renderHighHeatmap}}
}

