% Generated by roxygen2 (4.0.1): do not edit by hand
\name{renderHighcharts}
\alias{renderHighcharts}
\title{Render Highcharts}
\usage{
renderHighcharts(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that evaluates to a list that can
be converted to a JSON object for direct injection into a
highcharts function call.
See http://api.highcharts.com/highcharts for details on how
to specify highcharts arguments.}

\item{env}{}

\item{quoted}{}
}
\description{
Creates a Highcharts output
}
\seealso{
Other shinyhighcharts.elements: \code{\link{JSONify}};
  \code{\link{getHighchartsColors}};
  \code{\link{highchartsOutput}}
}

