% Generated by roxygen2 (4.0.1): do not edit by hand
\name{renderHighHeatmap}
\alias{renderHighHeatmap}
\title{Render Highcharts Heatmap}
\usage{
renderHighHeatmap(expr, args, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that evaluates to a matrix or dataframe object
with three columns.  The first column is the y axis value, the second
is the x axis value, and the third column is the color value for the
quadrant.}

\item{xTitle}{X axis label}

\item{yTitle}{Y axis label}

\item{xAxisCatagories}{A vector of x axis catagory labels, must be the
same length as the the nummber of unique x axis values in expr.}

\item{yAxisCatagories}{A vector of y axis catagory labels, must be the
same length as the the nummber of unique y axis values in expr.}

\item{invertXAxis}{Boolean, should the x axis be reversed.  Defaults to
FALSE.}

\item{invertYAxis}{Boolean, should the y axis be reversed.  Defaults to
TRUE.}

\item{env}{}

\item{quoted}{}
}
\description{
Creates a Highcharts heatmap output from a matrix of dataframe.
}
\seealso{
Other shinyhighcharts.elements: \code{\link{highchartsOutput}}
}

