% Generated by roxygen2 (4.0.1): do not edit by hand
\name{highchartsOutput}
\alias{highchartsOutput}
\title{Highcharts Output}
\usage{
highchartsOutput(inputId, width = "100\%", height = "400px",
  include = c("base", "more", "3d", "all", "canvas-tools", "data",
  "drilldown", "exporting", "funnel", "heatmap", "no-data", "solid-gauge"))
}
\arguments{
\item{inputId}{}

\item{width}{}

\item{height}{}

\item{include}{A vector indicating the highcharts source
 files to include in the page header.  See the Highcharts
 API reference for more information about the necessary
 source files for the chart types available.}
}
\description{
Render a Highcharts output within an application page.
}
\seealso{
Other shinyhighcharts.elements: \code{\link{JSONify}};
  \code{\link{getHighchartsColors}};
  \code{\link{renderHighcharts}}
}

